+++ test/test_elab.c
@@ -406,7 +406,7 @@ START_TEST(test_eval1)
 
    const error_t expect[] = {
       { 12, "array index -1 outside bounds 7 downto 0" },
-      { 16, "while evaluating call to \"=\"" },
+      { 16, "while evaluating call to FUNC" },
       { 16, "expression cannot be folded to an integer constant" },
       { -1, NULL }
    };
@@ -453,6 +453,19 @@ START_TEST(test_issue307)
 }
 END_TEST
 
+START_TEST(test_issue315)
+{
+   input_from_file(TESTDIR "/elab/issue315.vhd");
+
+   tree_t top = run_elab();
+   fail_if(top == NULL);
+
+   tree_t d2 = tree_decl(top, 2);
+   fail_unless(icmp(tree_ident(d2), ":issue315:info"));
+   fail_unless(tree_kind(tree_value(d2)) == T_AGGREGATE);
+}
+END_TEST
+
 int main(void)
 {
    Suite *s = suite_create("elab");
@@ -485,10 +498,11 @@ int main(void)
    tcase_add_test(tc, test_libbind3);
    tcase_add_test(tc, test_issue251);
    tcase_add_test(tc, test_jcore1);
-   tcase_add_exit_test(tc, test_eval1, EXIT_FAILURE);
+   tcase_add_test(tc, test_eval1);
    tcase_add_test(tc, test_issue305);
    tcase_add_test(tc, test_gbounds);
    tcase_add_test(tc, test_issue307);
+   tcase_add_test(tc, test_issue315);
    suite_add_tcase(s, tc);
 
    return nvc_run_test(s);