+++ HISTORY.md
@@ -3,6 +3,7 @@
 - Record aggregates can now be constant folded (#305)
 - The TCL shell is deprecated and will be removed in the next release
 - Compatible with LLVM 4.0
+- Functions returning records can now be constant folded (#315)
 
 ## 1.1 - 2016-11-05
 - Improvements to based literal parsing